+++ ts/base/hash-instance.ts
@@ -1,4 +1,5 @@
 import { BaseHashInput, inputToArray, IBaseHashOptions, defaultHashLength } from './hash-fn';
+import { IHashReader } from './hash-reader';
 
 /**
  * A blake3 hash. Quite similar to Node's crypto hashing.
@@ -15,8 +16,21 @@ export interface IHasher<T> {
 
   /**
    * Returns a digest of the hash.
+   *
+   * If `dispose: false` is given in the options, the hash will not
+   * automatically be disposed of, allowing you to continue updating
+   * it after obtaining the current reader.
    */
-  digest(options?: IBaseHashOptions): T;
+  digest(options?: IBaseHashOptions & { dispose?: boolean }): T;
+
+  /**
+   * Returns a {@link HashReader} for the current hash.
+   *
+   * If `dispose: false` is given in the options, the hash will not
+   * automatically be disposed of, allowing you to continue updating
+   * it after obtaining the current reader.
+   */
+  reader(options?: { dispose?: boolean }): IHashReader<T>;
 
   /**
    * Frees data associated with the hash. This *must* be called if
@@ -28,8 +42,9 @@ export interface IHasher<T> {
 /**
  * @hidden
  */
-export interface IInternalHash {
+export interface IInternalHash<Reader> {
   free(): void;
+  reader(): Reader;
   update(bytes: Uint8Array): void;
   digest(into: Uint8Array): void;
 }
@@ -37,21 +52,24 @@ export interface IInternalHash {
 /**
  * Base implementation of hashing.
  */
-export class BaseHash<T extends Uint8Array> implements IHasher<T> {
-  // these are covariant, but typing them better has a runtime overhead
-  private hash: IInternalHash | undefined;
-  private digested?: T;
+export class BaseHash<Binary extends Uint8Array, InternalReader, Reader extends IHashReader<Binary>>
+  implements IHasher<Binary> {
+  private hash: IInternalHash<InternalReader> | undefined;
 
-  constructor(implementation: IInternalHash, private readonly alloc: (length: number) => T) {
+  constructor(
+    implementation: IInternalHash<InternalReader>,
+    private readonly alloc: (length: number) => Binary,
+    private readonly getReader: (internal: InternalReader) => Reader,
+  ) {
     this.hash = implementation;
   }
 
   /**
    * @inheritdoc
    */
-  update(data: BaseHashInput): this {
+  public update(data: BaseHashInput): this {
     if (!this.hash) {
-      throw new Error('Cannot continue hashing after digest() or dispose() has been called');
+      throw new Error('Cannot continue updating hashing after dispose() has been called');
     }
 
     this.hash.update(inputToArray(data));
@@ -61,19 +79,38 @@ export class BaseHash<T extends Uint8Array> implements IHasher<T> {
   /**
    * @inheritdoc
    */
-  digest(options?: IBaseHashOptions): T {
-    if (this.digested) {
-      return this.digested;
-    }
-
+  public digest({
+    length = defaultHashLength,
+    dispose = true,
+  }: IBaseHashOptions & { dispose?: boolean } = {}): Binary {
     if (!this.hash) {
       throw new Error('Cannot call digest() after dipose() has been called');
     }
 
-    this.digested = this.alloc(options?.length ?? defaultHashLength);
-    this.hash.digest(this.digested);
-    this.hash.free();
-    return this.digested;
+    const digested = this.alloc(length);
+    this.hash.digest(digested);
+
+    if (dispose) {
+      this.dispose();
+    }
+
+    return digested;
+  }
+
+  /**
+   * @inheritdoc
+   */
+  public reader({ dispose = true }: { dispose?: boolean } = {}) {
+    if (!this.hash) {
+      throw new Error('Cannot call reader() after dipose() has been called');
+    }
+
+    const reader = this.getReader(this.hash.reader());
+    if (dispose) {
+      this.dispose();
+    }
+
+    return reader;
   }
 
   /**