+++ ts/node-native/index.ts
@@ -1,4 +1,5 @@
 export { HashInput } from '../node/hash-fn';
 export { hash } from './hash-fn';
+export * from '../node/hash-reader';
 export * from './hash-instance';
 export * from '../base';