+++ ts/node.test.ts
@@ -1,35 +1,28 @@
 import * as wasm from './node';
 import * as native from './node-native';
 import { expect } from 'chai';
-import { inputs } from './base/test-helpers';
+import { inputs, hello48 } from './base/test-helpers';
 import { ReadableStreamBuffer } from 'stream-buffers';
-
-const toHex = (arr: Uint8Array) => Buffer.from(arr).toString('hex');
+import { maxHashBytes } from './base/hash-reader';
 
 function suite({ hash, createHash }: typeof wasm | typeof native) {
   describe('encoding', () => {
     it('hashes a buffer', () => {
-      expect(hash(Buffer.from(inputs.large.input), { encoding: 'hex' })).to.equal(
-        inputs.large.hash,
-      );
+      expect(hash(Buffer.from(inputs.hello.input))).to.deep.equal(inputs.hello.hash);
     });
 
     it('hashes a string', () => {
-      expect(hash(inputs.large.input, { encoding: 'hex' })).to.equal(inputs.large.hash);
+      expect(hash(inputs.hello.input)).to.deep.equal(inputs.hello.hash);
     });
 
     it('hashes an arraybuffer', () => {
-      expect(hash(Buffer.from(inputs.large.input).buffer, { encoding: 'hex' })).to.equal(
-        inputs.large.hash,
-      );
+      const buf = Buffer.from(inputs.hello.input);
+      expect(hash(new Uint8Array(buf).buffer)).to.deep.equal(inputs.hello.hash);
     });
 
     it('customizes the output length', () => {
       expect(hash(inputs.hello.input, { length: 16 })).to.deep.equal(
-        Buffer.from(inputs.hello.hash, 'hex').slice(0, 16),
-      );
-      expect(hash(inputs.hello.input, { length: 16, encoding: 'hex' })).to.deep.equal(
-        inputs.hello.hash.slice(0, 32),
+        inputs.hello.hash.slice(0, 16),
       );
     });
   });
@@ -38,10 +31,10 @@ function suite({ hash, createHash }: typeof wasm | typeof native) {
     it('hash', () => {
       const hashA = hash('hello');
       const hashB = hash('goodbye');
-      expect(toHex(hashA)).to.equal(
+      expect(hashA.toString('hex')).to.equal(
         'ea8f163db38682925e4491c5e58d4bb3506ef8c14eb78a86e908c5624a67200f',
       );
-      expect(toHex(hashB)).to.equal(
+      expect(hashB.toString('hex')).to.equal(
         'f94a694227c5f31a07551908ad5fb252f5f0964030df5f2f200adedfae4d9b69',
       );
     });
@@ -56,10 +49,10 @@ function suite({ hash, createHash }: typeof wasm | typeof native) {
 
       const hashA = hasherA.digest();
       const hashB = hasherB.digest();
-      expect(toHex(hashA)).to.equal(
+      expect(hashA.toString('hex')).to.equal(
         'ea8f163db38682925e4491c5e58d4bb3506ef8c14eb78a86e908c5624a67200f',
       );
-      expect(toHex(hashB)).to.equal(
+      expect(hashB.toString('hex')).to.equal(
         'f94a694227c5f31a07551908ad5fb252f5f0964030df5f2f200adedfae4d9b69',
       );
     });
@@ -76,7 +69,7 @@ function suite({ hash, createHash }: typeof wasm | typeof native) {
       buffer.on('data', b => hash.update(b));
       buffer.on('end', () => {
         const actual = hash.digest();
-        expect(toHex(actual)).to.equal(inputs.large.hash);
+        expect(actual).to.deep.equal(inputs.large.hash);
         callback();
       });
     });
@@ -90,7 +83,7 @@ function suite({ hash, createHash }: typeof wasm | typeof native) {
         .pipe(createHash())
         .on('error', callback)
         .on('data', hash => {
-          expect(toHex(hash)).to.equal(inputs.large.hash);
+          expect(hash).to.deep.equal(inputs.large.hash);
           callback();
         });
     });
@@ -98,7 +91,64 @@ function suite({ hash, createHash }: typeof wasm | typeof native) {
     it('customizes the output length', () => {
       const hash = createHash();
       hash.update(inputs.hello.input);
-      expect(hash.digest('hex', { length: 16 })).to.equal(inputs.hello.hash.slice(0, 32));
+      expect(hash.digest('hex', { length: 16 })).to.equal(
+        inputs.hello.hash.slice(0, 16).toString('hex'),
+      );
+    });
+  });
+
+  describe('reader', () => {
+    let reader: wasm.NodeHashReader;
+    beforeEach(() => {
+      const hash = createHash();
+      hash.update(inputs.hello.input);
+      reader = hash.reader();
+    });
+
+    afterEach(() => reader.dispose());
+
+    it('implements toString()', () => {
+      expect(reader.toString('hex')).to.equal(inputs.hello.hash.toString('hex'));
+      reader.position = BigInt(42);
+      expect(reader.toString('hex')).to.equal(inputs.hello.hash.toString('hex'));
+    });
+
+    it('implements toBuffer()', () => {
+      expect(reader.toBuffer()).to.deep.equal(inputs.hello.hash);
+      reader.position = BigInt(42);
+      expect(reader.toBuffer()).to.deep.equal(inputs.hello.hash);
+    });
+
+    it('implements readInto() and advances', () => {
+      const actual = Buffer.alloc(32);
+      reader.readInto(actual.slice(0, 10));
+      reader.readInto(actual.slice(10));
+      expect(actual).to.deep.equal(inputs.hello.hash);
+      expect(reader.position).to.equal(BigInt(32));
+    });
+
+    it('implements read() and advances', () => {
+      const actual = reader.read(32);
+      expect(actual).to.deep.equal(inputs.hello.hash);
+      expect(reader.position).to.equal(BigInt(32));
+
+      const actualNext = reader.read(16);
+      expect(actualNext).to.deep.equal(hello48.slice(32));
+      expect(reader.position).to.equal(BigInt(48));
+    });
+
+    it('manually sets position', () => {
+      reader.position = BigInt(32);
+      const actual = reader.read(16);
+      expect(actual).to.deep.equal(hello48.slice(32));
+    });
+
+    it('throws if set out of range', () => {
+      expect(() => (reader.position = BigInt(-1))).to.throw(RangeError);
+      expect(() => (reader.position = BigInt('18446744073709551616'))).to.throw(RangeError);
+
+      reader.position = maxHashBytes - BigInt(1);
+      expect(() => reader.read(2)).to.throw(RangeError);
     });
   });
 }