+++ ts/node/hash-fn.ts
@@ -15,22 +15,11 @@ export const normalizeInput = (input: HashInput, encoding?: BufferEncoding): Uin
 /**
  * Returns a blake3 hash of the input, returning the binary hash data.
  */
-export function hash(input: HashInput, options?: IBaseHashOptions): Buffer;
-
-/**
- * Returns a blake3 hash of the input, returning the hash encoding with the
- * requested encoding.
- */
 export function hash(
   input: HashInput,
-  options: IBaseHashOptions & { encoding: BufferEncoding },
-): string;
-
-export function hash(
-  input: HashInput,
-  { encoding, length = defaultHashLength }: IBaseHashOptions & { encoding?: BufferEncoding } = {},
+  { length = defaultHashLength }: IBaseHashOptions = {},
 ): Buffer | string {
   const result = Buffer.alloc(length);
   rawHash(normalizeInput(input), result);
-  return encoding ? result.toString(encoding) : result;
+  return result;
 }