+++ ts/node/index.ts
@@ -1,3 +1,4 @@
 export { hash, HashInput } from './hash-fn';
 export * from './hash-instance';
+export * from './hash-reader';
 export * from '../base';